move ReadableIORef to util sublibrary
