+++ app/Swarm/App.hs
@@ -34,12 +34,12 @@ import Graphics.Vty.Platform.Unix qualified as VS
 import Swarm.Game.Failure (SystemFailure)
 import Swarm.Language.Pretty (prettyText)
 import Swarm.Log (LogSource (SystemLog), Severity (..))
-import Swarm.ReadableIORef (mkReadonly)
 import Swarm.TUI.Controller
 import Swarm.TUI.Model
 import Swarm.TUI.Model.StateUpdate
 import Swarm.TUI.Model.UI (uiAttrMap)
 import Swarm.TUI.View
+import Swarm.Util.ReadableIORef (mkReadonly)
 import Swarm.Version (getNewerReleaseVersion)
 import Swarm.Web
 import System.IO (stderr)
